Game accuracy calculation

+++ src/levelData.h
@@ -11,12 +11,12 @@ namespace synthrush {
 struct LevelData {
     std::string levelName;
     std::string audioFile;
-    std::vector<int> beats;
+    std::vector<float> beats;
 
     LevelData(json j) {
         levelName = j.at("levelName").get<std::string>();
         audioFile = j.at("audioFile").get<std::string>();
-        beats = j.at("beats").get<std::vector<int>>();
+        beats = j.at("beats").get<std::vector<float>>();
     }
 };
